Moved unused components and files to another folder
