+++ public/components/unusedComponents/typewriter.tsx
